backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
streamlit==1.32.0
qdrant-client==1.7.0
fastembed==0.2.2
groq==0.4.1
llama-index-core>=0.10.11
llama-index-embeddings-fastembed>=0.1.4
llama-index-llms-groq>=0.1.4
llama-index-vector-stores-qdrant>=0.1.3
bcrypt>=4.0.1